chore: update docs to use raw jsx instead of template literal
